Expose current RBR execution state in the SQL Thread state
Summary:
WebScaleSQL Feature: Row-Based Replication Improvements

Port the following patch from twitter to WebScaleSql

https://github.com/twitter/mysql/commit/3a1e4b5682d994ccefc919a33d9ac7d2adc

Expose current RBR execution state in the SQL Thread state

The execution of row-based events in a replication slave is not
reflected in the SQL Thread state, making it rather difficult to
identify long-running events. For example, the execution of a large
row-by-row delete event is not immediately identifiable as the
replication SQL thread state for most of its duration stays as
"Reading event from the relay log".

This change adds two new states to the replication slave SQL thread
that are used to highlight the event that the SQL thread is executing
and, for row-based events, to indicate how many rows have been
applied. These will generally look like:

        Executing Delete_rows event at position 100
        Handling record 50 of 100 for a Delete_rows event on table test.t1

Additionally, the outermost state of a thread state is now saved and
restored when inner states are set, so that the overall information
of what the thread is doing is not lost.

Changes during porting
1. Extend the functionality to MTS
2. thd_proc_info is replaced with THD_STAGE_INFO() in 5.6. Save thd->proc_info
   before THD_STAGE_INFO() and restore it back.

Test Plan:
mtr tests

Add new mtr test with MTS.

Reviewers: inaam-rana, darnaut, xiaobin.lin, weixiang.zhai

Reviewed By: weixiang.zhai

Subscribers: steaphan

Differential Revision: https://reviews.facebook.net/D31533
Differential Revision: https://reviews.facebook.net/D62169