Expose user certificate details to command line.
Summary:
This is a change to help the implementation of SSL based authentication efforts. Currently the certificate details are unavailable to the sys admin. The v3 extensions usually carry important information for identifying the user.

This change adds a new table to the information schema to expose auth details. It is organized by process ID like processlist table.

Test Plan: Run newly added test case

Reviewers: pengt, tianx, jtolmer, santoshb

Reviewed By: santoshb

Differential Revision: https://reviews.facebook.net/D53169