Fix incorrect results when defining a view with doc paths using the merge algorithm
Summary:
When selectin from a VIEW, the merge algorithm tries to rewrite the query to
use the original view definition. However, if the definition of the view has
a document path, the "merge" algorithm does NOT write the full document path
but rather, it only uses the document column name.

For example, consider
```
create view v1 (id) as
select doc.id from t1;

select id from v1;
```

The select query will be rewritten to:
```select doc from t1```
instead of
```select doc.id from t1```

This diff correctly fixes the merge algorithm to print the full document path
name in the appropriate places.

This diff disallows unresolved idents when selecting from a view. This means
that having a document path on top of a document path alias
is not supported for now.

Test Plan: Added more test cases for views in json.type_document_path

Reviewers: pengt

Reviewed By: pengt