Using Field_document as the field type instead of Field_varstring in temporal table for document path
Summary:
Currently, in temporal table, document path is treated as Field_varstring
to make it work in group by. But Doing like this can make the column
loosing all the information needed to be treated as a document. So it will
not work in temporal table and sub query very well. In this diff we change
the type of column of Field_varstring in temporal table when created to the
original type: Field_document.
By doing this, we fix the problem of sub query, order by, the limited size of
the column of temporal table(4096 currently).
But there are still  two problems remaining:
  -- one is that when a column has two aliases for different document path.
Because, currently the document path is stored in Field_document, so they
conflict with each other.
  -- the other one is that when we do group by, we still need to make it as Field_varstring.
Because, group by is done in MyISAM, and make it as Field_document will
make it not work correctly.

Test Plan: Jenkins

Reviewers: tianx, pengt

Reviewed By: pengt