Log detailed instructions to enable RocksDB if it's disabled
Summary:
This set of changes is a fix for https://github.com/facebook/mysql-5.6/issues/91. If RocksDB isn't specified as a storage engine then we'll log instructions about how to enable it.

Currently the message goes to `mysqld.?.err `file, but if we need it to be more prominently present then we can log directly to stdout as well. It's trivial to make that change.

Test Plan:
Execute the following basic test-cases:

  - Set up MySQL with the default engine (InnoDB) and observe that the newly introduced message is being present in `*.err` file.

  - Set up MySQL with RocksDB as a storage engine and observe that the message is //not// present in `*.err `file.

Reviewers: yoshinorim, MarkCallaghan

Reviewed By: MarkCallaghan

Subscribers: webscalesql-eng

Differential Revision: https://reviews.facebook.net/D52773