Handle NULL return (out of memory) in mysql_delete
Summary:
It is possible for substitute_for_best_equal_fields() to return NULL, but this was not being checked before dereferencing the value.  When a huge query-alloc-block-size value was specified (and we were running a release build) this function would attempt to allocate a new block and fail and then return NULL which then cause a segfault.

With this change in place the NULL return value is handled correctly and the client gets the out of memory error.

Test Plan: MTR - I attempted to write a test for this case, but when I caused an OOM at this location so many other things failed too that it wasn't feasible within MTR.  I did manually run the client and got this memory error and the server no longer crashes.

Reviewers: yoshinorim, hermanlee4

Reviewed By: hermanlee4

Subscribers: webscalesql-eng

Differential Revision: https://reviews.facebook.net/D53571