Fix the known runtime errors pointed out by ubsan
Summary:
As preparation for enabling a dedicated UndefinedBehaviorSanitizer (`ubsan`) build we need to fix all the known issues exposed by pre-commit MySQL tests so that we can have a clean baseline.

Fixes here are trivial: fix a casting issue and make sure that we'll shift the bits in the allowed range only. There are tricker (and more indecipherable) ways to fix them, but I chose readability over brevity.

Test Plan:
  # Produce a `ubsan` build.
  # Run `mysqltest.sh` and `mysqltest.sh --testset=RocksDB`.
  # Make sure that all the test cases pass.

Reviewers: MarkCallaghan, vasilep

Reviewed By: vasilep

Subscribers: jkedgar, webscalesql-eng

Differential Revision: https://reviews.facebook.net/D55779