Add option to exclude tables from gap lock check
Summary:
Similar to what was done in D49035, a new sysvar was added called `gap_lock_exceptions` that takes a comma separated list of regexes that specify which tables should skip gap lock detection. If it is skipped, then we do not raise errors, and we do not log to the gaplock log.

In order to do this, I've moved some logic from the collations check to `sql/handler.cc`. I've also made changes the mutex that protected the table list into a rwlock.

Fixes #232

Test Plan: mtr

Reviewers: yoshinorim, hermanlee4, jkedgar

Reviewed By: jkedgar

Subscribers: webscalesql-eng

Differential Revision: https://reviews.facebook.net/D56937
Differential Revision: https://reviews.facebook.net/D61299