Fixed conflicts on hard-coded port 14000
Summary:
Two of the tests were using a hard-coded port (14000).  This won't work if the tests get run simultaneously by two different jobs on the same test machine.  Refactored to use the master port + 1 instead.

The original idea had been to start a master and a slave and then shut down the slave and use it's port as the admin port of the master server.  This turned out to be overly complicated and I had failures trying to start the slave up again when I was done (I thought it would be wise to leave everything in the same state as the start of the test).  This method turns out to be simpler - though it assumes that the server's main port plus one is available.  This should be a valid assumption as the MTR framework allocates 10 ports for each thread running tests.

Test Plan: MTR

Reviewers: hermanlee4, pengt

Reviewed By: pengt

Subscribers: tianx, webscalesql-eng

Differential Revision: https://reviews.facebook.net/D58221