Support loose index range scans for low cardinality column predicates
Summary:
Suppose we have the following query:

  CREATE TABLE tbl (
    a int,
    b int,
    c int,
    primary key (a, b, c)
  )

  SELECT * FROM tbl WHERE a = 5 AND c > 100 ORDER BY a, b, c;

The current behaviour in MySQL is to do a range scan on the whole range where a = 5, and then filter using the c > 100 predicate. However, we can be smarter than this.
Instead, we do a loose index scan by scanning over distinct (a, b), and then doing a subrange scan on that distinct prefix for c > 100.

The implementation is heavily inspired by the existing loose index scan done for GROUP BY and DISTINCT queries. A new TABLE_READ_PLAN and QUICK_SELECT_I that implements this strategy was added. We check before creating the TABLE_READ_PLAN whether this plan is applicable or not. After that, we execute the loose index scan in the QUICK_LOOSE_RANGE_SELECT::get_next method.

Test Plan: mtr

Reviewers: tianx, spetrunia

Reviewed By: spetrunia

Subscribers: webscalesql-eng

Differential Revision: https://reviews.facebook.net/D59877
Differential Revision: https://reviews.facebook.net/D64833