innobackupex is reporting some invalid information for the master log offset and name
Summary: The innobackupex utility prints out some information related to the master log offset and name, but this information doesn't appear to be used anywhere.  We have at least one replicaset that has invalid information at this location and when the file name is generated it is not ascii, which confuses the Python script that is interpreting the output from the utility.  Change the innobackupex utility to not generate this output when the other data indicates that the data may be bad.

Test Plan: MTR

Reviewers: hermanlee4

Reviewed By: hermanlee4

Subscribers: webscalesql-eng

Differential Revision: https://reviews.facebook.net/D61077