[libmysql] Do not set socket to blocking mode in nonblocking functions
Summary:
We are making blocking calls after queries in the async
client. It looks like it was part of a guard to set the socket to
blocking mode to ensure existing functions work, and to deal with some
weirdness when socket modes would change.
This should be safe to remove but I'll do a better code review and tests
to ensure we aren't hitting weird issues

Test Plan: unittests

Reviewers: mung, anca

Reviewed By: mung, anca

Subscribers: webscalesql-eng

Differential Revision: https://reviews.facebook.net/D64071
Differential Revision: https://reviews.facebook.net/D64443