Fixed flaky rpl.rpl_gtid_failover test
Summary: The test queried a global variable and in one case the result set contained two rows.  The SELECT statement used did not specify an order and in at least one case the order that the rows came back in was different.  Added an ORDER BY to make sure the order is always consistent.

Test Plan: MTR

Reviewers: hermanlee4, santoshb, tianx

Reviewed By: tianx

Subscribers: webscalesql-eng

Differential Revision: https://reviews.facebook.net/D64419
Differential Revision: https://reviews.facebook.net/D64863
Differential Revision: https://reviews.facebook.net/D65493