Fix xtrabackup tests
Summary: A number of xtrabackup tests used the xb_load_data.sh script.  This inserted 10,000 rows, then spawned a background test to insert another 10,000 rows while simultaneously running xtrabackup.  Later the tests started up replication and synced the slave with the master.  At this point they select the count of rows in the table and assumed that the background thread had finished and the number would be 20,000.  We've had some cases where this wasn't the case - it would appear that the machine was busy enough that the background shell script was not finishing before replication had been started and synced up.  Adding code to have the script wait for the background script (using a temporary file as an indicator).

Test Plan: MTR

Reviewers: hermanlee4, tianx

Reviewed By: tianx

Subscribers: webscalesql-eng

Differential Revision: https://reviews.facebook.net/D64467