Fix deadlock detection tests
Summary:
Checking "Waiting for row locks" is not reliable because that message will display even if the thread is not blocked. One way to work around this is to use the newly created rocksdb_locks table to see if all the locks we expect the transaction to take has been taken. The better way is to use rocksdb_trx once we have that.
Closes https://github.com/facebook/mysql-5.6/pull/393

Reviewed By: alxyang

Differential Revision: D4128101

Pulled By: lth

fbshipit-source-id: 4eda293