Delete crash safe index file if index file exists
Summary:
Crash safe index file is created when any change is made to the index
file. The index file is first copied over to the crash safe one and then stuff
is added/removed from the crash safe index file. Finally, the old index file is
deleted and crash safe index file is renamed to the index file. There might be a
scenario where mysql crashes after stuff is changed in the crash safe file
but before the old index file is deleted. In this case we should delete the
crash safe index file at restart. If we don't delete it subsequent changes to
the index file can corrupt it because it can be copied to the old crash safe
file which already has some data in it.
Closes https://github.com/facebook/mysql-5.6/pull/433

Reviewed By: santoshbanda

Differential Revision: D4198274

Pulled By: abhinav04sharma

fbshipit-source-id: a6f8993