Fix problem with Oracle for bug 78777/21974225 not working on normal build
Summary: When Oracle fixed bug 78777/21974225 they didn't seem to test it without performance schema.  The 'init' stage is being set when the network header is read, but without performance schema there is no access to the THD at that point.  Added a call to set the 'init' stage back in sql_parse where the command is read from the network.

Test Plan: MTR

Reviewers:

Subscribers:

Tasks:

Blame Revision: